requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
fastapi==0.112.1
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.3.0
mmh3==4.1.0
nltk==3.8.1
numpy==1.26.4
openai==1.37.1
packaging==24.1
pandas==2.2.2
pinecone==5.0.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-dotenv==1.0.1
pytz==2024.1
rank-bm25==0.2.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soylemma==0.2.0
starlette==0.38.2
tqdm==4.66.4
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
wget==3.2
